requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1

# Document processing
pypdf==4.3.1
pandas==2.2.2
openpyxl==3.1.5

# Embeddings and ML
sentence-transformers==3.1.1
torch==2.4.1
transformers==4.44.2

# Retrieval
rank-bm25==0.2.2
psycopg2-binary==2.9.9
pgvector==0.3.2
sqlalchemy==2.0.35

# LLM
google-generativeai==0.8.1
langchain==0.3.1
langchain-google-genai==2.0.1

# Agents
langgraph==0.2.28

# Evaluation
ragas==0.4.2
datasets==4.8.4

# Observability
langfuse==2.53.3

# Search
duckduckgo-search==6.3.0

# Utilities
httpx==0.27.2
tenacity==8.5.0
apscheduler==3.10.4
python-multipart==0.0.12